requirements.txt
django==2.0.2
Pillow 
pylint
pylint-django
django-bower
xlwt
xlrd
pandas
django-mathfilters
djangorestframework
psycopg2==2.8.5
djangorestframework==3.11.2
django-manifest-loader